+++ ModelProject/src/main/java/org/liferayasif/backend/model/PersonKin.java
@@ -20,11 +20,11 @@ public class PersonKin implements Serializable {
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
 	@Column(name="PERSON_KIN_ID")
-	private Long id;
+	private Integer id;
 	
 	
 	@Column(name="PERSON_ID")
-	private Long personId ;
+	private Integer personId ;
 	
 	@Column(name="NAME")
 	private String  name;
@@ -49,19 +49,19 @@ public class PersonKin implements Serializable {
 	@Transient
 	public List<PersonContact> personContactList;
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
 
-	public Long getPersonId() {
+	public Integer getPersonId() {
 		return personId;
 	}
 
-	public void setPersonId(Long personId) {
+	public void setPersonId(Integer personId) {
 		this.personId = personId;
 	}
